tmkms/requirements.txt
PyNaCl==1.5.0
cryptography==44.0.2
